Updating direct IO write checksum verifies
Previously, we used zfs_vdev_direct_write_verify_cnt to set the number
of Direct I/O writes that were issued to a top-level vdev before a
checksum verify would be performed on a Direct I/O write. However, this
required grabbing the VDEV vdev_stat_lock to increment a counter. This
was not very efficient.

Brian Behlendorf suggested that we instead use a percentage of Direct
I/O writes to a top-level vdev before doing a checksum verify check.
This allows us to avoid the locking and can accomplish the same thing.

I udpated the module parameter to be zfs_vdev_direct_write_verify_pct
which is now interpreted as the percentage of Direct I/O writes to a
top-level VDEV before we do the checksum verification. By default, I
have set this to 2 percent for Linux. We still don't need this on
FreeBSD as it can place user pages under write protection so it is set
to 0 percent by default for FreeBSD.

As part of this, I updated the man pages to reflect the change in the
module parameter and it's new meaning as well as updated the ZTS test
dio_write_verify to use the new parameter properly.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>