WIP: See if we can do hole punch reads
Attempting to see if we can actually do reads with Direct I/O that do
not require us to read the whole record but just the small amount
requested. This does imply that a checksum verify is never run on the
returned data. The idea that is being experimented with here is to see
if IOPs to a NVMe SSD can be driven up by allowing for sub-blocksize
reads to take place.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>